pubsys: move region_from_string to common aws module
